# Optimize wrapping with SGR